+++ samples/astarte_app/src/main.c
@@ -282,9 +282,9 @@ static void device_rx_thread_entry_point(void *arg1, void *arg2, void *arg3)
 
     LOG_INF("End of loop, disconnection imminent."); // NOLINT
 
-    res = astarte_device_disconnect(device);
+    res = astarte_device_disconnect(device, K_SECONDS(10));
     if (res != ASTARTE_RESULT_OK) {
-        LOG_ERR("Astarte device disconnection failure."); // NOLINT
+        LOG_ERR("Astarte device disconnection failure %s.", astarte_result_to_name(res)); // NOLINT
         return;
     }