+++ CHANGELOG.md
@@ -10,6 +10,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - Unreleased
+### Added
+- Function `astarte_device_force_disconnect` forces the device disconnection from an Astarte
+  instance discarding any pending QoS 1/2 messages.
+
+### Changed
+- The `astarte_device_disconnect` function waits for all pending QoS 1/2 messages to be correctly
+  processed with a configurable a timeout.
+
 ## [0.7.2] - 2024-10-23
 ### Changed
 - All the library samples have been merged into a single sample called `astarte_app`.